+++ node.h
@@ -19,6 +19,6 @@ int _node_comparator_memaddr(void *value1, void *value2);
  * A hash value function for nodes using the memory address of the data as the
  * hash value.
  */
-unsigned int _node_hashvalue_memaddr(void *value);
+unsigned long _node_hashvalue_memaddr(void *value);
 
 #endif  /* NODE_H */
